requirements.txt
cupy==9.6.0
einops==0.4.1
meshzoo==0.11.1
mmflow==0.5.2
numpy==1.21.5
omegaconf==2.2.3
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
pandas==1.5.0
Pillow==10.0.0
PyYAML==6.0
rawpy==0.17.2
scikit_image==0.19.3
timm==0.4.12
torch==1.12.1
torchinfo==1.7.1
torchvision==0.13.1
tqdm==4.64.1
hydra-core==1.2.0
